test/requirements.txt
googlemaps
osmnx
networkx
shapely
geopandas
pysolar
pandas
requests
numpy
python-dotenv